+++ order/src/main/java/com/toanyone/order/application/service/OrderService.java
@@ -6,7 +6,7 @@ import com.toanyone.order.application.dto.service.request.OrderCancelServiceDto;
 import com.toanyone.order.application.dto.service.request.OrderCreateServiceDto;
 import com.toanyone.order.application.dto.service.request.OrderFindAllCondition;
 import com.toanyone.order.application.dto.service.request.OrderSearchCondition;
-import com.toanyone.order.infrastructure.kafka.OrderKafkaProducer;
+import com.toanyone.order.application.port.out.OrderMessageProducer;
 import com.toanyone.order.infrastructure.mapper.ItemRequestMapper;
 import com.toanyone.order.infrastructure.converter.OrderMessageConverter;
 import com.toanyone.order.common.dto.CursorPage;
@@ -16,9 +16,9 @@ import com.toanyone.order.domain.model.Order;
 import com.toanyone.order.domain.model.OrderItem;
 import com.toanyone.order.domain.repository.OrderItemRepository;
 import com.toanyone.order.domain.repository.OrderRepository;
-import com.toanyone.order.message.DeliveryRequestMessage;
-import com.toanyone.order.message.PaymentCancelMessage;
-import com.toanyone.order.message.PaymentRequestMessage;
+import com.toanyone.order.application.dto.message.DeliveryRequestMessage;
+import com.toanyone.order.application.dto.message.PaymentCancelMessage;
+import com.toanyone.order.application.dto.message.PaymentRequestMessage;
 import com.toanyone.order.presentation.dto.response.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
