+++ order/src/main/java/com/toanyone/order/common/config/RedisConfig.java
@@ -1,6 +1,5 @@
 package com.toanyone.order.common.config;
 
-import com.toanyone.order.message.DeliveryRequestMessage;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
