+++ order/src/main/java/com/toanyone/order/infrastructure/converter/OrderMessageConverter.java
@@ -1,8 +1,8 @@
 package com.toanyone.order.infrastructure.converter;
 
 import com.toanyone.order.application.dto.service.request.OrderCreateServiceDto;
-import com.toanyone.order.message.DeliveryRequestMessage;
-import com.toanyone.order.message.PaymentRequestMessage;
+import com.toanyone.order.application.dto.message.DeliveryRequestMessage;
+import com.toanyone.order.application.dto.message.PaymentRequestMessage;
 import org.springframework.stereotype.Component;
 
 import java.util.stream.Collectors;
