+++ order/src/test/java/com/toanyone/order/application/OrderServiceTest.java
@@ -6,7 +6,7 @@ import com.toanyone.order.application.dto.service.request.ItemRestoreRequestDto;
 import com.toanyone.order.application.dto.service.request.ItemValidationRequestDto;
 import com.toanyone.order.application.dto.service.response.HubFindResponseDto;
 import com.toanyone.order.application.dto.service.response.StoreFindResponseDto;
-import com.toanyone.order.infrastructure.kafka.OrderKafkaProducer;
+import com.toanyone.order.infrastructure.adapter.kafka.OrderKafkaProducer;
 import com.toanyone.order.infrastructure.mapper.ItemRequestMapper;
 import com.toanyone.order.infrastructure.converter.OrderMessageConverter;
 import com.toanyone.order.application.service.*;
@@ -16,9 +16,9 @@ import com.toanyone.order.domain.model.Order;
 import com.toanyone.order.domain.model.OrderItem;
 import com.toanyone.order.domain.repository.OrderItemRepository;
 import com.toanyone.order.domain.repository.OrderRepository;
-import com.toanyone.order.message.DeliveryRequestMessage;
-import com.toanyone.order.message.PaymentCancelMessage;
-import com.toanyone.order.message.PaymentRequestMessage;
+import com.toanyone.order.application.dto.message.DeliveryRequestMessage;
+import com.toanyone.order.application.dto.message.PaymentCancelMessage;
+import com.toanyone.order.application.dto.message.PaymentRequestMessage;
 import com.toanyone.order.presentation.dto.response.OrderCancelResponseDto;
 import com.toanyone.order.presentation.dto.response.OrderCreateResponseDto;
 import org.junit.jupiter.api.*;
